Moved all enriching logic to processing/enriching.py. Implemented function to enrich centres DataFrames with weather data
